+++ src/commands/init.ts
@@ -1,18 +1,20 @@
-import { Command } from '@oclif/command';
+import { Command, flags } from '@oclif/command';
 import { args as Parser } from '@oclif/parser';
 import chalk from 'chalk';
 import { prompt, Question } from 'inquirer';
 
+import { readConfig, writeConfig } from '../config/fs';
 import { DEFAULT_CONFIG_PATH } from '../constants';
 import { Environment } from '../environment';
 import { make as makeExample } from '../example';
 import { quietFlag, WithQuietFlag } from '../flags/quiet';
-import { AwsConfig, ProjectConfig, writeConfig } from '../projectConfig';
+import { stageFlag, WithStageFlag } from '../flags/stage';
 
 // Defined to name Args interface properties as constants.
 const AWS_ACCESS = 'awsAccess';
 const AWS_SECRET = 'awsSecret';
 const ROOT_PATH = 'rootPath';
+const STAGES = 'stages';
 
 /**
  * Defines the interface of the answers received from `inquirer`
@@ -21,7 +23,11 @@ interface Answers {
   awsAccess: string;
   awsSecret: string;
   rootPath: string;
+  stages: string;
 }
+
+type AnswerKey = keyof Answers;
+type ShouldQuestion = Record<AnswerKey, boolean>;
 
 /**
  * Defines the information received as positional arguments. It must be at least
@@ -30,7 +36,7 @@ interface Answers {
 interface Args extends Partial<Answers> {}
 
 /** Defines the information received as flags. */
-interface Flags extends WithQuietFlag {} // tslint:disable-line no-empty-interface
+interface Flags extends WithQuietFlag, WithStageFlag {} // tslint:disable-line no-empty-interface
 
 export class Init extends Command {
   static description = 'Create the configuration files for your project.';
@@ -38,7 +44,7 @@ export class Init extends Command {
   static examples = [
     makeExample([
       `# Create configuration with given parameters.`,
-      `$ ssmenv init / FOO bar`,
+      `$ ssmenv init --stage=production / FOO bar`,
       `Configuration written to .ssmenv/public.json and .ssmenv/private.json.`,
       `* Recommend adding .ssmenv/public.json to source control.`,
       `* Recommend ignoring .ssmenv/private.json in source control.`,
@@ -50,6 +56,7 @@ export class Init extends Command {
       `? AWS Access Key ID`,
       `? AWS Secret Access Key`,
       `? Root Path (/)`,
+      `? Comma Separated Stages (development,production,staging,test)`,
       `Configuration written to .ssmenv/public.json and .ssmenv/private.json.`,
       `* Recommend adding .ssmenv/public.json to source control.`,
       `* Recommend ignoring .ssmenv/private.json in source control.`,
@@ -58,6 +65,7 @@ export class Init extends Command {
 
   static flags = {
     quiet: quietFlag,
+    stage: stageFlag,
   };
 
   static args: Parser.IArg[] = [
@@ -74,6 +82,47 @@ export class Init extends Command {
       name: AWS_SECRET,
     },
   ];
+
+  /**
+   * Uses `Environment#validatePathPart` to check for errors but catches any
+   * thrown to translate them into a `string` message.
+   * @param input to be validated.
+   * @return `true` if `input` is valid, a `string` message if `input` is not
+   *    valid.
+   */
+  static isValidPathPart(input: string) {
+    try {
+      Environment.validatePathPart('Stage', input);
+      return true;
+    } catch (err) {
+      if (err instanceof Error) {
+        return err.message;
+      } else {
+        throw err;
+      }
+    }
+  }
+
+  /**
+   * Uses `isValidPathPart` to check for errors in the comma separated list of
+   * path parts in `input.
+   * @param input of comma separated path parts to validate.
+   * @return `true` if `input` is valid, a `string` message if `input` is not
+   *    valid.
+   */
+  static isValidPathParts(input: string) {
+    const parts = input.split(',');
+    const isValidResults = parts
+      .map(Init.isValidPathPart)
+      .filter(result => typeof result === 'string');
+    if (isValidResults.length === 0) {
+      // no parts were invalid so `input` is valid.
+      return true;
+    } else {
+      // return a single string
+      return isValidResults.join(' ');
+    }
+  }
 
   /**
    * Uses `Environment#validateRootPath` to check for errors but catches any
@@ -96,58 +145,114 @@ export class Init extends Command {
   }
 
   /**
+   * Create default values for Questions based on current configuration, given
+   * `args` and given `flags`.
+   * @param args parsed for to the current execution.
+   * @param flags parsed for the current execution.
+   * @return values to uses as defaults if a new value is not provided.
+   */
+  private static async defaultAnswers(
+    args: Args,
+    flags: Flags
+  ): Promise<Partial<Answers>> {
+    const currentConfig = await readConfig(DEFAULT_CONFIG_PATH);
+    return {
+      awsAccess: args.awsAccess || currentConfig.accessKeyId,
+      awsSecret: args.awsSecret || currentConfig.secretAccessKey,
+      rootPath: args.rootPath || currentConfig.rootPath || '/',
+      stages:
+        (flags.stage && flags.stage.join(',')) ||
+        (currentConfig.stages && currentConfig.stages.join(',')) ||
+        'development,production,staging,test',
+    };
+  }
+
+  /**
+   * Create mapping indicating if a configuration value must be prompted.
+   * @param args parsed for to the current execution.
+   * @param flags parsed for the current execution.
+   * @return a mapping of keys from `Answers` to boolean values, `true`
+   *    indicating the question should be prompted.
+   */
+  private static isMissingAnswer(args: Args, flags: Flags): ShouldQuestion {
+    return {
+      awsAccess: args.awsAccess === undefined,
+      awsSecret: args.awsSecret === undefined,
+      rootPath:
+        args.rootPath === undefined ||
+        Init.isValidRootPath(args.rootPath) !== true,
+      stages:
+        flags.stage === undefined ||
+        flags.stage.length === 0 ||
+        Init.isValidPathParts(flags.stage.join(',')) !== true,
+    };
+  }
+
+  /**
    * Get the `Question` instances to prompt for based on the `Args` provided.
    * @param args which may provide answers to `Question`s already.
+   * @param flags which may provide answers to `Question`s already.
    * @returns the set of `Question`s to with their `when` condition set based on
    *    `args`.
    */
-  private static questions(args: Args): Question[] {
+  private static async questions(
+    args: Args,
+    flags: Flags
+  ): Promise<Question[]> {
+    const isMissingValue = Init.isMissingAnswer(args, flags);
+    const defaults = await Init.defaultAnswers(args, flags);
     return [
       {
+        default: defaults.awsAccess,
         message: 'AWS Access Key ID',
         name: AWS_ACCESS,
         type: 'input',
-        when: args.awsAccess === undefined,
+        when: isMissingValue.awsAccess,
       },
       {
+        default: defaults.awsSecret,
         message: 'AWS Secret Access Key',
         name: AWS_SECRET,
         type: 'input',
-        when: args.awsSecret === undefined,
+        when: isMissingValue.awsSecret,
       },
       {
-        default: '/',
+        default: defaults.rootPath,
         message: 'Root Path',
         name: ROOT_PATH,
         type: 'input',
         validate: Init.isValidRootPath,
-        when:
-          args.rootPath === undefined ||
-          Init.isValidRootPath(args.rootPath) !== true,
+        when: isMissingValue.rootPath,
+      },
+      {
+        default: defaults.stages,
+        message: 'Comma Separated Stages',
+        name: STAGES,
+        type: 'input',
+        validate: Init.isValidPathParts,
+        when: isMissingValue.stages,
       },
     ];
   }
 
   async run() {
     const { args, flags } = this.parse<Flags, Args>(Init);
-    const answers = await prompt<Answers>(Init.questions(args));
+    const questions = await Init.questions(args, flags);
+    const answers = await prompt<Answers>(questions);
     const accessKeyId = args.awsAccess || answers.awsAccess;
-    const secretAccessKey = args.awsSecret || answers.awsSecret;
     const rootPath = args.rootPath || answers.rootPath;
+    const secretAccessKey = args.awsSecret || answers.awsSecret;
+    // Despite what the inferred signature of `flags` indicates `stage` can be undefined
+    const stages = flags.stage || answers.stages.split(',');
     // If checks didn't exit then we have valid values
-    const projectConfig: ProjectConfig = {
-      rootPath,
-    };
-    const awsConfig: AwsConfig = {
+    const config = {
       accessKeyId,
+      rootPath,
       secretAccessKey,
+      stages,
     };
 
-    const paths = await writeConfig(
-      awsConfig,
-      projectConfig,
-      DEFAULT_CONFIG_PATH
-    );
+    const paths = await writeConfig(config, DEFAULT_CONFIG_PATH);
     const keyword = chalk.keyword('green');
     const [awsPath, projectPath] = paths.map(v => keyword(v));
     const stdout = flags.quiet