+++ CHANGELOG.md
@@ -1,15 +1,44 @@
-<a name="0.3.0"></a>
+# Changelog
 
-# 0.3.0 (https://github.com/oursiberia/ssmenv/compare/v0.2.0...v0.3.0) (2018-04-09)
+All notable changes to this project will be documented in this file.
 
-### Bug Fixes
+The format is based on [Keep a Changelog][kac] and this project adheres to
+[Semantic Versioning][semver].
 
-    * commands: Create explicit *Index commands (a3d861d (https://github.com/oursiberia/ssmenv/commit/a3d861d)), closes #7 (https://github.com/oursiberia/ssmenv/issues/7)
-    * env: Create EnvIndex command (61302e4 (https://github.com/oursiberia/ssmenv/commit/61302e4)), closes #7 (https://github.com/oursiberia/ssmenv/issues/7)
-    * var: Create VarIndex command (1661263 (https://github.com/oursiberia/ssmenv/commit/1661263)), closes #7 (https://github.com/oursiberia/ssmenv/issues/7)
+[kac]: http://keepachangelog.com/en/1.0.0/
+[semver]: http://semver.org/spec/v2.0.0.html
 
-### Features
+# Unreleased
 
-    * env:list: Create env:list command (da1aaa8 (https://github.com/oursiberia/ssmenv/commit/da1aaa8)), closes #8 (https://github.com/oursiberia/ssmenv/issues/8)
-    * init: Add --quiet flag to init command (cf9464e (https://github.com/oursiberia/ssmenv/commit/cf9464e))
-    * var:del: Create var:del command (9afb91d (https://github.com/oursiberia/ssmenv/commit/9afb91d))
+## Added
+
+* `init` command takes multiple stages through the `--stage` flag.
+* `init` command prompts for stages during initialization if they are not
+  provided.
+* `init` command writes the stages to the `public.json` configuration file.
+* `var:del`, `var:set` and `var:tag` commands will now operate on multiple
+  stages through the `--stage` flag.
+* `var:del`, `var:set` and `var:tag` commands will a add a stage provided by
+  the `--stage` flag to the `public.json` configuration file if it is not
+  already present.
+
+## Changed
+
+* `var:del`, `var:set` and `var:tag` commands no longer take a positional
+  `STAGE` argument.
+
+# 0.3.0 ([since v0.2.0](https://github.com/oursiberia/ssmenv/compare/v0.2.0...v0.3.0)) (2018-04-09)
+
+## Added
+
+* Create `env:list` command to list all paths recognized as environments in
+  AWS SSM Parameter Store. Closes
+  [#8](https://github.com/oursiberia/ssmenv/issues/8)
+* Add `--quiet` flag to `init` command to suppress informational output.
+* Create `var:del` command which deletes a variable from the stage or
+  environment indicated by the positional stage argumnet.
+
+## Fixed
+
+* Create explicit command for `ssmenv env` and `ssmenv var`. Closes
+  [#7](https://github.com/oursiberia/ssmenv/issues/7)